Refactoring Error Messages

+++ app/Http/Controllers/UserController.php
@@ -3,9 +3,6 @@
 namespace App\Http\Controllers;
 
 use App\Comment;
-use App\Http\Requests\NewPasswordRequest;
-use App\Http\Requests\UserRequest;
-use App\Http\Requests\UserUpdateRequest;
 use App\Session;
 use App\User;
 use App\Video;
@@ -19,6 +16,7 @@ use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
@@ -38,9 +36,9 @@ class UserController extends Controller
             $user = User::query()->where('username', 'LIKE', $username)->with('record')->firstOrFail();
         } catch (\Exception $exception) {
             return response([
-                'message' => 'Usuario no Encontrado',
-                'error_message' => $exception->getMessage()
-            ], 422);
+                'from' => 'Info Login',
+                'error_message' => 'Credenciales Incorrectas'
+            ], 401);
         }
 
         $expire_time = $user->record->reset_password['password_expired_at'];
@@ -55,10 +53,9 @@ class UserController extends Controller
                     'iat' => now()->unix(),
                     'nbf' => now()->addMillisecond()->unix(),
                     'exp' => now()->addDays(2)->unix(),
-                    'user' => Auth::user()->channel
+                    'channel' => Auth::user()->channel
                 );
                 $encoded = JWT::encode($payload, env('APP_KEY'), 'HS512');
-                $decoded = JWT::decode($encoded, env('APP_KEY'), array('HS512'));
 
                 $refresh = JWT::encode(array(
                     'sub' => Auth::id(),
@@ -75,8 +72,9 @@ class UserController extends Controller
                 $new_session->save();
 
                 return response([
-                    'auth:message' => 'User Authenticated',
-                    'auth:user' => $decoded,
+                    'from' => 'Info Sesión',
+                    'message' => 'Sesión Iniciada con Éxito',
+                    'auth_user' => Auth::user()->channel,
                     'X-Authentication-JWT' => $encoded,
                     'X-Refresh-JWT' => $refresh,
                     'X-Encode-ID' => Crypt::encrypt(Auth::id())
@@ -84,13 +82,14 @@ class UserController extends Controller
 
             } else {
                 return response([
-                    'message' => 'Usuario no Autenticado',
-                ], 422);
+                    'from' => 'Info Login',
+                    'error_message' => 'Credenciales Incorrectas'
+                ], 401);
             }
         } else {
             return response([
-                'message' => 'Su Contraseña Expiró',
-                'expired_at' => $expire_time
+                'from' => 'Info Login',
+                'error_message' => 'Su Contraseña ha Expirado'
             ], 401);
         }
     }
@@ -108,14 +107,15 @@ class UserController extends Controller
             JWT::decode($jwt_refresh, env('APP_KEY'), array('HS256'));
         } catch (\Exception $exception) {
             return response([
-                'message' => 'Sesión Comprometida',
-                'error_message' => $exception->getMessage()
+                'from' => 'Info Sesión',
+                'error_message' => 'Sesión Comprometida'
             ], 401);
         }
         Session::query()->where('jwt_refresh', 'LIKE', $jwt_refresh)->delete();
         Auth::logout();
         return response([
-            'message' => 'User Logout Successfully'
+            'from' => 'Info Sesión',
+            'message' => 'Sesión Cerrada con Éxito'
         ], 200);
     }
 
@@ -132,8 +132,8 @@ class UserController extends Controller
             $jwt_refresh_decoded = JWT::decode($jwt_refresh, env('APP_KEY'), array('HS256'));
         } catch (\Exception $exception) {
             return response([
-                'message' => 'Sesión Comprometida',
-                'error_message' => $exception->getMessage()
+                'from' => 'Info Sesión',
+                'error_message' => 'Sesión Comprometida'
             ], 401);
         }
 
@@ -143,7 +143,8 @@ class UserController extends Controller
         } else {
             $session->delete();
             return response([
-                'message' => 'Seguridad Comprometida, Sesión Cerrada'
+                'from' => 'Info Sesión',
+                'error_message' => 'Seguridad Comprometida, Sesión Cerrada'
             ], 401);
         }
 
@@ -152,11 +153,10 @@ class UserController extends Controller
             'iat' => now()->unix(),
             'nbf' => now()->addMillisecond()->unix(),
             'exp' => now()->addDays(2)->unix(),
-            'user' => Auth::user()->channel
+            'channel' => Auth::user()->channel
         );
 
         $encoded = JWT::encode($payload, env('APP_KEY'), 'HS512');
-        $decoded = JWT::decode($encoded, env('APP_KEY'), array('HS512'));
 
         $new_jwt_refresh = JWT::encode(array(
             'sub' => Auth::id(),
@@ -170,50 +170,74 @@ class UserController extends Controller
         ]);
 
         return response([
-            'message' => 'User Login Successfully',
-            'auth:message' => 'User Authenticated',
-            'auth:user' => $decoded,
+            'from' => 'Info Usuario',
+            'message' => 'Sesión Recuperada',
+            'auth_user' => Auth::user()->channel,
             'X-Authentication-JWT' => $encoded,
             'X-Refresh-JWT' => $new_jwt_refresh,
             'X-Encode-ID' => Crypt::encrypt(Auth::id())
-        ], 200);
+        ], 200)->withHeaders([
+            'X-Authentication-JWT' => $encoded,
+            'X-Refresh-JWT' => $new_jwt_refresh,
+            'X-Encode-ID' => Crypt::encrypt(Auth::id())
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param UserRequest $request
+     * @param Request $request
      * @return Response
      */
-    public function store(UserRequest $request)
+    public function store(Request $request)
     {
-        $fromUserRequest = $request->all();
-        $newUser = new User($fromUserRequest);
-        $newUser['password'] = Hash::make($newUser['password']);
-        try {
-            $newUser->save();
-        } catch (Exception $e) {
+        $validator = Validator::make($request->all(), [
+            'fullname' => 'required|string',
+            'username' => 'required|unique:users|min:4|string',
+            'email' => 'required|email:rfc,strict,spoof,filter|unique:users',
+            'password' => 'required|min:8|confirmed'
+        ], [], [
+            'fullname' => 'nombre completo',
+            'username' => 'usuario',
+            'email' => 'correo electrónico',
+            'password' => 'contraseña',
+            'password_confirmation' => 'confirmación de contraseña'
+        ]);
+
+        if ($validator->fails()) {
             return response([
-                'message' => 'Usuario no Guardado',
-                'error_message' => $e->getMessage(),
+                'from' => 'Info Usuario',
+                'errors' => $validator->errors()->all()
             ], 422);
         }
+
+        $newUser = new User($request->all());
+        $newUser['password'] = Hash::make($newUser['password']);
+        $newUser->save();
+
         return response([
-            'message' => 'User Stored',
-            'user_id' => $newUser->refresh()->id
+            'from' => 'Info Usuario',
+            'message' => 'Su Cuenta ha sido creada'
         ], 201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param User $user
+     * @param $user
      * @return Response
      */
-    public function show(User $user)
+    public function show($user)
     {
+        try {
+            $user = User::query()->findOrFail($user);
+        } catch (\Exception $exception) {
+            return response([
+                'from' => 'Info Usuario',
+                'error_message' => 'El usuario solicitado no existe o no está disponible'
+            ], 404);
+        }
         return response([
-            'message' => 'User Found',
             'channel' => $user->channel
         ], 200);
     }
@@ -221,12 +245,39 @@ class UserController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param UserUpdateRequest $request
-     * @param User $user
+     * @param Request $request
+     * @param $user
      * @return Response
      */
-    public function update(UserUpdateRequest $request, User $user)
+    public function update(Request $request, $user)
     {
+        try {
+            $user = User::query()->findOrFail($user);
+        } catch (\Exception $exception) {
+            return response([
+                'from' => 'Info Usuario',
+                'error_message' => 'El usuario solicitado no existe o no está disponible'
+            ], 404);
+        }
+
+        $validator = Validator::make($request->all(), [
+            'fullname' => 'nullable|string',
+            'username' => ['nullable', 'min:4', 'string', Rule::unique('users')->ignore($user->id)],
+            'email' => ['nullable', 'email:rfc,strict,spoof,filter', Rule::unique('users')->ignore($user->id)],
+            'avatar' => 'nullable|file|image|max:10240'
+        ], [], [
+            'fullname' => 'nombre completo',
+            'username' => 'usuario',
+            'email' => 'correo electrónico',
+        ]);
+
+        if ($validator->fails()) {
+            return response([
+                'from' => 'Info Usuario',
+                'errors' => $validator->errors()->all()
+            ], 422);
+        }
+
         if (request()->hasFile('avatar')) {
             Storage::delete('public/uploads/channel-' . $user->channel->id . '/avatar/' . $user->channel->avatar['name']);
             $fileAvatar = request()->file('avatar');
@@ -237,7 +288,6 @@ class UserController extends Controller
         }
         $user->update($request->all());
         return response([
-            'message' => 'User Updated',
             'user' => $user->refresh()->channel
         ], 200);
     }
@@ -245,46 +295,61 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param User $user
+     * @param $user
      * @return Response
      * @throws Exception
      */
-    public function destroy(User $user)
+    public function destroy($user)
     {
-        Storage::deleteDirectory('public/uploads/channel-' . $user->channel->id);
         try {
-            Session::query()->where('user_id', 'LIKE', $user->id)->delete();
-            Video::query()->where('channel_id', 'LIKE', $user->channel->id)->delete();
-            Comment::query()->where('user_id', 'LIKE', $user->id)->delete();
-            $user->channel()->delete();
-            $user->record()->delete();
-            $user->delete();
-        } catch (Exception $e) {
+            $user = User::query()->findOrFail($user);
+        } catch (\Exception $exception) {
             return response([
-                'message' => 'Usuario no Eliminado',
-                'error_message' => $e->getMessage(),
-            ], 422);
+                'from' => 'Info Usuario',
+                'error_message' => 'El usuario solicitado no existe o no está disponible'
+            ], 404);
         }
+
+        Storage::deleteDirectory('public/uploads/channel-' . $user->channel->id);
+
+        Session::query()->where('user_id', 'LIKE', $user->id)->sharedLock()->delete();
+        Video::query()->where('channel_id', 'LIKE', $user->channel->id)->sharedLock()->delete();
+        Comment::query()->where('user_id', 'LIKE', $user->id)->sharedLock()->delete();
+        $user->channel()->delete();
+        $user->record()->delete();
+        $user->delete();
+
         return response([
-            'message' => 'User Deleted'
+            'from' => 'Info Usuario',
+            'message' => 'Usuario Eliminado con Éxito'
         ], 200);
     }
 
     /**
      * Change the current password from a user
-     * @param NewPasswordRequest $request
+     * @param Request $request
      * @param Faker $faker
      * @return Response
      */
-    public function newPassword(NewPasswordRequest $request, Faker $faker)
+    public function newPassword(Request $request, Faker $faker)
     {
-        $fromRequest = $request->all();
-        if (!Hash::check($fromRequest['current_password'], Auth::user()->getAuthPassword())) {
+        $validator = Validator::make($request->all(), [
+            'current_password' => 'required|min:8|password',
+            'new_password' => 'required|min:8|confirmed'
+        ], [], [
+            'current_password' => 'contraseña actual',
+            'new_password' => 'contraseña nueva',
+            'new_password_confirmation' => 'confirmación de contraseña nueva',
+        ]);
+
+        if ($validator->fails()) {
             return response([
-                'message' => 'La Contraseña no es Correcta'
+                'from' => 'Info Usuario',
+                'errors' => $validator->errors()->all()
             ], 422);
         }
-        $user = User::query()->findOrFail(Auth::id());
+
+        $user = User::query()->find(Auth::id());
         $user->record()->update([
             'reset_password' => [
                 'secret_list' => $user->record->reset_password['secret_list'],
@@ -293,10 +358,11 @@ class UserController extends Controller
             ]
         ]);
         $user->update([
-            'password' => Hash::make($fromRequest['new_password'])
+            'password' => Hash::make($request->get('current_password'))
         ]);
         return response([
-            'message' => 'Contraseña Actualizada'
+            'from' => 'Info Usuario',
+            'message' => 'Contraseña Actualizada. Inicie Sesión para actualizar los Cambios'
         ], 200);
     }
 
@@ -313,12 +379,11 @@ class UserController extends Controller
             JWT::decode($jwt_temp, env('APP_KEY'), array('HS512'));
         } catch (\Exception $exception) {
             return response([
-                'message' => 'Petición no Autorizada',
-                'error_message' => $exception->getMessage()
+                'from' => 'Info Seguridad',
+                'error_message' => 'Petición no Autorizada'
             ], 401);
         }
         return response([
-            'message' => 'Secret List',
             'secret_list' => [
                 $faker->state,
                 $faker->state,
@@ -336,14 +401,32 @@ class UserController extends Controller
 
     /**
      * Check if new User is OK
-     * @param UserRequest $request
+     * @param Request $request
      * @return Response
      */
-    public function checkNewUser(UserRequest $request)
+    public function checkNewUser(Request $request)
     {
-        return response([
-            'message' => 'Ready for Store'
-        ], 200);
+        $validator = Validator::make($request->all(), [
+            'fullname' => 'required|string',
+            'username' => 'required|unique:users|min:4|string',
+            'email' => 'required|email:rfc,strict,spoof,filter|unique:users',
+            'password' => 'required|min:8|confirmed'
+        ], [], [
+            'fullname' => 'nombre completo',
+            'username' => 'usuario',
+            'email' => 'correo electrónico',
+            'password' => 'contraseña',
+            'password_confirmation' => 'confirmación de contraseña'
+        ]);
+
+        if ($validator->fails()) {
+            return response([
+                'from' => 'Info Usuario',
+                'errors' => $validator->errors()->all()
+            ], 422);
+        }
+
+        return response([], 200);
     }
 
     /**
@@ -355,29 +438,28 @@ class UserController extends Controller
     public function resetPassword(Request $request, Faker $faker)
     {
         $validator = Validator::make($request->all(), [
-            'email' => 'required|email:rfc,strict,spoof,filter',
+            'email' => 'required|email:rfc,strict,spoof,filter|exists:users',
             'array_numbers' => 'required',
             'first_word' => 'required',
             'second_word' => 'required',
             'third_word' => 'required'
+        ], [], [
+            'email' => 'correo electrónico',
+            'array_numbers' => 'palabra secreta',
+            'first_word' => 'palabra secreta',
+            'second_word' => 'palabra secreta',
+            'third_word' => 'palabra secreta'
         ]);
 
         if ($validator->fails()) {
             return response([
-                'message' => 'Info Incorrecta',
-                'error_message' => $validator->errors()
+                'from' => 'Info',
+                'errors' => $validator->errors()->all()
             ], 422);
         } else {
             $email = $request->get('email');
 
-            try {
-                $user = User::query()->where('email', 'LIKE', $email)->with('record')->firstOrFail();
-            } catch (\Exception $exception) {
-                return response([
-                    'message' => 'Usuario no Encontrado',
-                    'error_message' => $exception->getMessage()
-                ], 422);
-            }
+            $user = User::query()->where('email', 'LIKE', $email)->with('record')->first();
 
             $array_numbers = $request->get('array_numbers');
             $secret_list = $user->record->reset_password['secret_list'];
@@ -400,13 +482,13 @@ class UserController extends Controller
                 ]);
 
                 return response([
-                    'message' => 'Password Recovered',
                     'new_password' => $password
                 ], 200);
 
             } else {
                 return response([
-                    'message' => 'Info Incorrecta, inténtalo otra vez'
+                    'from' => 'Info',
+                    'errors' => ['Palabras Incorrectas, inténtalo otra vez']
                 ], 422);
             }
         }
